app/requirements.txt
fastapi
uvicorn
python-telegram-bot==20.7
openai
python-dotenv
